More style cleanup
